+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+# Changed
+
+- [TD-4536] Support rule implementation with multiple populations
+
 ## [4.41.1] 2022-04-04
 
 # Added